# dist/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Twitter Connector",
    "description": "All things you can use in this API"
  },
  "basePath": "/twitter",
  "tags" : [
    {
      "name": "Twitter",
      "description": "API for connecting Twitter"
    }
  ],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/path/to/method": {
      "post": {
        "tags": ["Twitter"],
        "description": "what this API does",
        "operationId": "for the URL",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Job Parameters",
            "in": "body",
            "description": "Job to add to the database",
            "required": true,
            "schema": {
              "$ref": "#/definitions/create"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "create response",
            "schema": {
              "$ref": "#/definitions/response"
            }
          }
        }
      }
    },
    "/path/to/method/{someVariable}": {
      "delete":{
        "tags":["Twitter"],
        "description":"Deletes a something",
        "operationId":"delete",
        "produces": [
          "application/json"
        ],
        "parameters":[
          {
            "name": "someVariable",
            "in": "path",
            "description": "ID to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "delete response",
            "schema":{
              "$ref":"#/definitions/deleteRes"
            }
          }
        }
      },
      "get":{
        "tags": [ "Twitter"],
        "description": "Returns the status of something",
        "operationId": "findById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "ID",
            "in": "path",
            "description": "ID to fetch",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "response",
            "schema": {
              "$ref": "#/definitions/response1"
            }
          }
        }
      }
    },
    "/": {
      "get":{
        "tags": [ "Twitter"],
        "description": "Home page",
        "operationId": "",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "response"
          }
        }
      }
    },
    "/index": {
      "get":{
        "tags": [ "Twitter"],
        "description": "index page",
        "operationId": "",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "response"
          }
        }
      }
    },
    "/tweet": {
      "post":{
        "tags": [ "Twitter"],
        "description": "Tweet something to an account",
        "operationId": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "status",
            "in": "body",
            "description": "Tweet",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "response"
          }
        }
      }
    },
    "/searchTweet": {
      "post":{
        "tags": [ "Twitter"],
        "description": "Seacrh tweets with any keywords",
        "operationId": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "q",
            "in": "body",
            "description": "keywords",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "response"
          }
        }
      }
    },
    "/searchTweetNearby": {
      "post":{
        "tags": [ "Twitter"],
        "description": "Seacrh tweets with latitude and longgitude",
        "operationId": "",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "lat",
            "in": "body",
            "description": "latitude",
            "required": true,
            "type": "number"
          },
          {
            "name": "lng",
            "in": "body",
            "description": "longitude",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "response"
          }
        }
      }
    }
  },
  "definitions": {
    "create": {
      "properties": {
        "eventId": {
          "required":true,
          "type": "string",
          "example": "mathmaticial"
        },
        "numberRequired": {
          "required":true,
          "type":"integer",
          "format": "int64",
          "example": 500
        }
      }
    },
    "response1": {
      "properties": {
        "jobID": {
          "required":true,
          "type": "string",
          "example": "23423456543"
        },
        "request": {
          "required":true,
          "type": "schema",
          "example": {
            "eventId":"mathmaticial",
            "numberRequired": 500
          }
        }
      }
    },
    "response": {
      "properties": {
        "jobID": {
          "required": true,
          "type" : "string",
          "example" : "12321432423"
        }
      }
    },
    "delete": {
      "properties": {
        "jobID": {
          "required": true,
          "type" : "string",
          "example" : "234344234234"
        },
        "status" : {
          "required": true,
          "type": "string",
          "example": "deleted"
        }
      }
    },
    "400error" : {
      "properties": {
        "message": {
          "type": "string",
          "example" : "400 Bad Request"
        }
      }
    }
  }
}